pyproject.toml
[project]
name = "multi-agent-routing-rag"
version = "0.1.0"
description = "Skeleton for a LangChain multi-agent orchestrator routing queries to HR/TECH RAG specialists."
requires-python = ">=3.10"
dependencies = [
  "langchain>=1.0.0",
  "langchain-chroma>=1.1.0",
  "langchain-groq>=1.1.2",
  "langchain-huggingface>=1.2.0",
  "langchain-openai>=1.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.1",
  "sentence-transformers>=5.2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
